move data channel logic to new file and deprecate message packet
