+++ client/src/redux/actionType.js
@@ -1,6 +1,5 @@
 export const ADD_PEER = "ADD_PEER";
 export const DELETE_PEER = "DELETE_PEER";
-export const ADD_MESSAGE = "ADD_MESSAGE";
 export const UPDATE_MY_UUID = "UPDATE_MY_UUID";
 export const UPDATE_PEER_UUID = "UPDATE_PEER_UUID";
 export const ADD_FILES = "ADD_FILES";
