+++ client/src/utils/peerConnection.js
@@ -13,8 +13,7 @@ import {
     messageResponseSharingData,
 } from "./dataSchema/PeerMessageData";
 import { createMessage } from "./message";
-import { createPeerMessage, parsePeerMessage } from "./peerMessage";
-import { createMessagePacket } from "./messagePacket";
+import { createPeerMessage } from "./peerMessage";
 import { generateFingerPrint } from "./commonUtil";
 import {
     sendMessageToServer,
@@ -22,26 +21,17 @@ import {
     deleteLeavedPeers,
     createMyUserInfo,
     getMyUUID,
-    addMessagePacket,
-    transferFileToPeer,
-    writePeerChunk,
     writeSystemMessage,
-    abortDownloadFile,
     connectToPeer,
     selectTableUsersMyself,
-    upsertTableUser,
     deleteTableUserByID,
-    upsertTableSharingData,
-    selectTableSharingDataByID,
-    patchTableSharingDataByID,
-    notifySharingData,
-    patchTableUsersByID,
-    selectTableSharingDataByUserID,
-    notifySharingDataToPeer,
 } from "./localApi";
-import { EventSendUserInfo } from "./dataSchema/LocalDropEventData";
-import LocalDropEvent from "./LocalDropEvent";
 import { initGlueSQL } from "./database/database";
+import {
+    handleDataChannelMessage,
+    registerDataChannelEventOnClose,
+    registerDataChannelEventOnOpen,
+} from "./dataChannel";
 
 function createPeerConnection(uuid) {
     console.log("createPeerConnection called", uuid);
@@ -57,36 +47,9 @@ function createPeerConnection(uuid) {
         }
     );
 
-    console.log("dataChannel:", dataChannel);
-
     dataChannel.binaryType = "arraybuffer";
-    dataChannel.onopen = async (event) => {
-        console.log("dataChanel onopen called", uuid);
-
-        handleDataChannelStatusChange(event, uuid);
-
-        const e = new LocalDropEvent(
-            CLIENT_EVENT_TYPE.SEND_USER_INFO,
-            new EventSendUserInfo({ uuid })
-        );
-
-        (await peerConnectionManager).dispatchEvent(e);
-
-        const e2 = new LocalDropEvent(
-            CLIENT_EVENT_TYPE.REQUEST_DATA_LIST,
-            null
-        );
-
-        (await peerConnectionManager).dispatchEvent(e2);
-    };
-
-    dataChannel.onclose = (event) => {
-        console.log(`data channel ${uuid} closed`);
-        writeSystemMessage("dataChannel closed");
-        handleDataChannelStatusChange(event, uuid);
-
-        abortDownloadFile(uuid);
-    };
+    registerDataChannelEventOnOpen(peerConnectionManager, dataChannel, uuid);
+    registerDataChannelEventOnClose(dataChannel, uuid);
 
     dataChannel.onmessage = (event) => {
         // console.log('[Message from DataChannel]: ', event);
@@ -160,144 +123,6 @@ function createPeerConnection(uuid) {
     };
 
     return { peerConnection, dataChannel };
-}
-
-async function handleDataChannelMessage(event, uuid) {
-    if (typeof event.data !== "string") {
-        const arrayBuffer = event.data;
-
-        await writePeerChunk(arrayBuffer, uuid);
-
-        return;
-    }
-
-    const message = parsePeerMessage(event.data);
-
-    console.log(
-        `[peer ${uuid}]: handleDataChannelMessage, message is `,
-        message
-    );
-
-    const { messageType, data } = message;
-
-    if (messageType === PEER_MESSAGE_TYPE.USER_INFO) {
-        // upsert to database
-        const { message } = data;
-
-        console.log("upsert user!", data);
-
-        if (!!message && Object.keys(message).length === 3) {
-            await upsertTableUser(message);
-        }
-
-        return;
-    }
-
-    if (messageType === PEER_MESSAGE_TYPE.REQUEST_DATA_LIST) {
-        const sharingDataList = await selectTableSharingDataByUserID(getMyUUID());
-
-        // make message and send back with data
-        for (const sharingData of sharingDataList) {
-            await notifySharingDataToPeer(sharingData, uuid);
-        }
-
-        return;
-    }
-
-    if (messageType === PEER_MESSAGE_TYPE.RESPONE_DATA_LIST) {
-        const { sharingData } = data;
-        await upsertTableSharingData({ sharingData });
-
-        return;
-    }
-
-    if (messageType === PEER_MESSAGE_TYPE.UPLOAD_LINK) {
-        const { sharingData } = data;
-        await upsertTableSharingData({ sharingData });
-
-        return;
-    }
-
-    if (messageType === PEER_MESSAGE_TYPE.UPDATE_USER) {
-        const { user } = data;
-        const { id: userID, name, profile } = user;
-        await patchTableUsersByID({ name, profile }, userID);
-
-        return;
-    }
-
-    if (messageType === PEER_MESSAGE_TYPE.TEXT) {
-        const messagePacket = createMessagePacket({
-            source: uuid,
-            destination: getMyUUID(),
-            data,
-            messageType,
-        });
-
-        console.log(
-            "[handleDataChannelMessage]: messagePacket is ",
-            messagePacket
-        );
-        addMessagePacket(messagePacket);
-
-        return;
-    }
-
-    if (messageType === PEER_MESSAGE_TYPE.FILE) {
-        const messagePacket = createMessagePacket({
-            source: uuid,
-            destination: getMyUUID(),
-            data,
-            messageType,
-        });
-
-        console.log(
-            "[handleDataChannelMessage]: messagePacket is ",
-            messagePacket
-        );
-        addMessagePacket(messagePacket);
-
-        return;
-    }
-
-    if (messageType === PEER_MESSAGE_TYPE.DOWNLOAD) {
-        const { fingerprint } = data;
-        // transfer file
-        transferFileToPeer(fingerprint, uuid);
-
-        // update download count, notify to peers
-        const sharingData = await selectTableSharingDataByID(fingerprint);
-        if (sharingData) {
-            const updated = await patchTableSharingDataByID(
-                { statusCount: sharingData.status_count + 1 },
-                fingerprint
-            );
-            await notifySharingData(updated);
-        }
-
-        return;
-    }
-
-    if (messageType === PEER_MESSAGE_TYPE.ERROR) {
-        const { message } = data;
-
-        writeSystemMessage(message);
-        return;
-    }
-}
-
-function handleDataChannelStatusChange(event, uuid) {
-    console.log(
-        `[peer ${uuid}]: handleDataChannelStatusChange, event is `,
-        event
-    );
-    const { type: eventType } = event;
-
-    // TODO: Delete dataChannel in peerConnectionManager
-    if (eventType === "close") {
-        const systemMessage = `[peer #${uuid}]: dataChannel closed `;
-        writeSystemMessage(systemMessage);
-    }
 }
 
 // TODO: Handle error later
@@ -433,16 +258,6 @@ function addClientEventTypeEventListener(peerConnectionManager) {
             }
 
             dataChannel.send(peerMessage);
-
-            // TODO(young): In orka, storing message packet is not necessary. Remove this when refactoring
-            // const messagePacket = createMessagePacket({
-            //     source: getMyUUID(),
-            //     destination: toUUID,
-            //     data,
-            //     messageType: PEER_MESSAGE_TYPE.TEXT,
-            // });
-
-            // addMessagePacket(messagePacket);
         }
     );
 
@@ -476,7 +291,7 @@ function addClientEventTypeEventListener(peerConnectionManager) {
                 console.log(
                     "CLIENT_EVENT_TYPE.REQUEST_DATA_LIST, message is ",
                     peerMessage
-                );                
+                );
 
                 if (dataChannel.readyState !== "open") {
                     console.log(
@@ -609,14 +424,14 @@ function addClientEventTypeEventListener(peerConnectionManager) {
                     peerMessage
                 );
 
-                dataChannel.send(peerMessage);
-
                 if (dataChannel.readyState !== "open") {
                     console.log(
                         `dataChannel (${dataChannel.readyState}) not opened for uuid: ${uuid}!, click the peer again!`
                     );
                     continue;
                 }
+
+                dataChannel.send(peerMessage);
             }
         }
     );
@@ -658,15 +473,6 @@ function addClientEventTypeEventListener(peerConnectionManager) {
             }
 
             dataChannel.send(peerMessage);
-
-            const messagePacket = createMessagePacket({
-                source: getMyUUID(),
-                destination: uuid,
-                data,
-                messageType: PEER_MESSAGE_TYPE.TEXT,
-            });
-
-            addMessagePacket(messagePacket);
         }
     );
 
@@ -703,15 +509,6 @@ function addClientEventTypeEventListener(peerConnectionManager) {
             }
 
             dataChannel.send(peerMessage);
-
-            const messagePacket = createMessagePacket({
-                source: getMyUUID(),
-                destination: uuid,
-                data,
-                messageType: PEER_MESSAGE_TYPE.FILE,
-            });
-
-            addMessagePacket(messagePacket);
         }
     );
